adds functioning etag to both queued spots and when trying to publish
fixes SPOT-767
refs SPOT-751

had to use a session variable when there is a failed If-Match from the
etags. the failed_json data is set to a session variable and loaded in
the diff template, but it gets immediately deleted to not rack up
unnecessary session variables. wish i could find a way to do it wihtout
a session variable :(