errors that are thrown for fields users cant edit are sent to the error template
fixes SPOT-777

before, errors were added to the QueuedSpace error field and remained there
until they were fixed. if an error got added to the field for something that
wasnt the users fault, specifically some etag or id error (something the user
cant fix), the space would get locked up and uneditable. this sends any of
those types of errors to the error template and doesnt save them in the error
field.
it also better handles how the error template displays information with varying
amounts of different types of data